ENV1/requirements.txt
django
django-mako
tweepy
python-instagram
Image
